+++ app/app.go
@@ -551,6 +551,7 @@ func New(
 		[]vm.PrecompiledContract{
 			cronosprecompiles.NewBankContract(app.BankKeeper, appCodec),
 			cronosprecompiles.NewRelayerContract(app.IBCKeeper, appCodec),
+			cronosprecompiles.NewIcaContract(&app.ICAAuthKeeper, appCodec),
 		},
 		allKeys,
 	)