+++ integration_tests/cosmoscli.py
@@ -1063,10 +1063,10 @@ class CosmosCLI:
         )
 
     def query_gravity_params(self):
-        return json.loads(self.raw("query", "gravity", "params", home=self.data_dir))
+        return self.query_params("gravity")
 
-    def query_evm_params(self):
-        return json.loads(self.raw("query", "evm", "params", home=self.data_dir))
+    def query_params(self, module="cronos"):
+        return json.loads(self.raw("query", module, "params", home=self.data_dir))
 
     def query_signer_set_txs(self):
         return json.loads(
@@ -1276,7 +1276,7 @@ class CosmosCLI:
             rsp = self.event_query_tx_for(rsp["txhash"])
         return rsp
 
-    def icaauth_submit_tx(self, connid, tx, **kwargs):
+    def icaauth_submit_tx(self, connid, tx, timeout_duration="1h", **kwargs):
         default_kwargs = {
             "home": self.data_dir,
             "node": self.node_rpc,
@@ -1290,6 +1290,8 @@ class CosmosCLI:
                 "submit-tx",
                 connid,
                 tx,
+                "--timeout-duration" if timeout_duration else None,
+                timeout_duration if timeout_duration else None,
                 "-y",
                 **(default_kwargs | kwargs),
             )
@@ -1314,7 +1316,7 @@ class CosmosCLI:
             )
         )
 
-    def ica_generate_packet_data(self, tx, memo=None, **kwargs):
+    def ica_generate_packet_data(self, tx, memo=None, encoding="proto3", **kwargs):
         return json.loads(
             self.raw(
                 "tx",
@@ -1322,7 +1324,8 @@ class CosmosCLI:
                 "host",
                 "generate-packet-data",
                 tx,
-                "--memo" if memo else None,
+                memo=memo,
+                encoding=encoding,
                 home=self.data_dir,
                 **kwargs,
             )
@@ -1584,17 +1587,6 @@ class CosmosCLI:
                 "-y",
                 home=self.data_dir,
                 **kwargs,
-            )
-        )
-
-    def query_params(self):
-        "query cronos params"
-        return json.loads(
-            self.raw(
-                "query",
-                "cronos",
-                "params",
-                home=self.data_dir,
             )
         )