+++ integration_tests/ibc_utils.py
@@ -597,6 +597,21 @@ def funds_ica(cli, adr):
     rsp = cli.transfer("signer2", adr, "1cro", gas_prices="1000000basecro")
     assert rsp["code"] == 0, rsp["raw_log"]
     wait_for_new_blocks(cli, 1)
-
+    amt = 100000000
     # check if the funds are received in interchain account
-    assert cli.balance(adr, denom="basecro") == 100000000
+    assert cli.balance(adr, denom="basecro") == amt
+    return amt
+
+
+def assert_channel_open_init(rsp):
+    assert rsp["code"] == 0, rsp["raw_log"]
+    port_id, channel_id = next(
+        (
+            evt["attributes"][0]["value"],
+            evt["attributes"][1]["value"],
+        )
+        for evt in rsp["events"]
+        if evt["type"] == "channel_open_init"
+    )
+    print("port-id", port_id, "channel-id", channel_id)
+    return port_id, channel_id