+++ integration_tests/test_ibc_rly.py
@@ -4,10 +4,8 @@ import signal
 import subprocess
 
 import pytest
-from eth_utils import abi, keccak, to_checksum_address
+from eth_utils import keccak, to_checksum_address
 from pystarport import cluster
-from web3._utils.contracts import abi_to_signature, find_matching_event_abi
-from web3._utils.events import get_event_data
 from web3.datastructures import AttributeDict
 
 from .ibc_utils import (
@@ -24,14 +22,17 @@ from .utils import (
     CONTRACT_ABIS,
     bech32_to_eth,
     eth_to_bech32,
+    get_logs_since,
+    get_method_map,
+    get_topic_data,
     module_address,
     wait_for_fn,
     wait_for_new_blocks,
 )
 
 CONTRACT = "0x0000000000000000000000000000000000000065"
-method_map = None
-contract_info = None
+contract_info = json.loads(CONTRACT_ABIS["IRelayerModule"].read_text())
+method_map = get_method_map(contract_info)
 cronos_signer2 = ADDRS["signer2"]
 src_amount = 10
 src_denom = "basecro"
@@ -79,41 +80,6 @@ def rly_transfer(ibc):
         f"--home {str(path)}"
     )
     subprocess.run(cmd, check=True, shell=True)
-
-
-def get_method_map():
-    global contract_info
-    if contract_info is None:
-        contract_info = json.loads(CONTRACT_ABIS["IRelayerModule"].read_text())
-    global method_map
-    if method_map is None:
-        method_map = {}
-        for item in contract_info:
-            event_abi = find_matching_event_abi(contract_info, item["name"])
-            signature = abi_to_signature(event_abi)
-            key = f"0x{abi.event_signature_to_log_topic(signature).hex()}"
-            method_map[key] = signature
-    return method_map, contract_info
-
-
-def get_topic_data(w3, log):
-    method_map, info = get_method_map()
-    method = method_map[log.topics[0].hex()]
-    name = method.split("(")[0]
-    event_abi = find_matching_event_abi(info, name)
-    event_data = get_event_data(w3.codec, event_abi, log)
-    return name, event_data.args
-
-
-def get_logs(w3, start):
-    end = w3.eth.get_block_number()
-    return w3.eth.get_logs(
-        {
-            "fromBlock": start,
-            "toBlock": end,
-            "address": [CONTRACT],
-        }
-    )
 
 
 def coin_received(receiver, amt, denom):
@@ -165,13 +131,6 @@ def burn(burner, amt, denom):
     return {
         "burner": burner,
         "amount": [(amt, denom)],
-    }
-
-
-def update_client():
-    return {
-        "clientId": keccak(text="07-tendermint-0"),
-        "clientType": keccak(text="07-tendermint"),
     }
 
 
@@ -245,13 +204,12 @@ def test_ibc(ibc):
 
     wait_for_fn("balance change", check_balance_change)
     assert old_dst_balance + dst_amount == new_dst_balance
-    logs = get_logs(w3, start)
+    logs = get_logs_since(w3, CONTRACT, start)
     relayer0 = ibc.chainmain.cosmos_cli().address("relayer")
     relayer = to_checksum_address(bech32_to_eth(relayer0))
     cronos_addr = module_address("cronos")
     transfer_addr = module_address("transfer")
     expected = [
-        update_client(),
         recv_packet(relayer0, cronos_signer2, src_amount, src_denom),
         denom_trace(denom),
         *send_from_module_to_acc(transfer_addr, cronos_signer2, src_amount, denom),
@@ -260,9 +218,8 @@ def test_ibc(ibc):
         *send_from_module_to_acc(cronos_addr, cronos_signer2, dst_amount, dst_denom),
         write_ack(relayer0, cronos_signer2, src_amount, src_denom),
     ]
-
     for i, log in enumerate(logs):
-        method_name, args = get_topic_data(w3, log)
+        method_name, args = get_topic_data(w3, method_map, contract_info, log)
         assert args == AttributeDict(expected[i]), [i, method_name]
 
 
@@ -278,7 +235,7 @@ def test_ibc_incentivized_transfer(ibc):
     wait_for_new_blocks(cli, 1)
     start = w3.eth.get_block_number()
     amount = ibc_incentivized_transfer(ibc)
-    logs = get_logs(w3, start)
+    logs = get_logs_since(w3, CONTRACT, start)
     fee_denom = "ibcfee"
     fee = f"{src_amount}{fee_denom}"
     transfer_denom = "transfer/channel-0/basetcro"
@@ -288,7 +245,6 @@ def test_ibc_incentivized_transfer(ibc):
     feeibc_addr = module_address("feeibc")
     escrow = get_escrow_address(cli, channel)
     expected = [
-        update_client(),
         acknowledge_packet(),
         distribute_fee(src_relayer, fee),
         *send_coins(feeibc_addr, src_relayer, src_amount, fee_denom),
@@ -297,7 +253,6 @@ def test_ibc_incentivized_transfer(ibc):
         distribute_fee(cronos_signer2, fee),
         *send_coins(feeibc_addr, cronos_signer2, src_amount, fee_denom),
         fungible(checksum_dst_adr, cronos_signer2, amount, dst_denom),
-        update_client(),
         recv_packet(dst_adr, cronos_signer2, amount, transfer_denom),
         *send_coins(escrow, cronos_signer2, amount, dst_denom),
         fungible(cronos_signer2, checksum_dst_adr, amount, transfer_denom),
@@ -305,7 +260,7 @@ def test_ibc_incentivized_transfer(ibc):
     ]
     assert len(logs) == len(expected)
     for i, log in enumerate(logs):
-        method_name, args = get_topic_data(w3, log)
+        method_name, args = get_topic_data(w3, method_map, contract_info, log)
         assert args == AttributeDict(expected[i]), [i, method_name]
 
 
@@ -315,10 +270,8 @@ def get_transfer_source_tokens_topics(dst_adr, amount, contract, escrow):
     cronos_denom = f"cronos{contract}"
     transfer_cronos_denom = f"transfer/{channel}/{cronos_denom}"
     return [
-        update_client(),
         acknowledge_packet(),
         fungible(checksum_dst_adr, ADDRS["validator"], amount, cronos_denom),
-        update_client(),
         recv_packet(dst_adr, cronos_signer2, amount, transfer_cronos_denom),
         *send_coins(escrow, cronos_signer2, amount, cronos_denom),
         fungible(cronos_signer2, checksum_dst_adr, amount, transfer_cronos_denom),
@@ -335,13 +288,13 @@ def test_cronos_transfer_source_tokens(ibc):
     w3 = ibc.cronos.w3
     start = w3.eth.get_block_number()
     amount, contract = cronos_transfer_source_tokens(ibc)
-    logs = get_logs(w3, start)
+    logs = get_logs_since(w3, CONTRACT, start)
     escrow = get_escrow_address(cli, channel)
     dst_adr = ibc.chainmain.cosmos_cli().address("signer2")
     expected = get_transfer_source_tokens_topics(dst_adr, amount, contract, escrow)
     assert len(logs) == len(expected)
     for i, log in enumerate(logs):
-        method_name, args = get_topic_data(w3, log)
+        method_name, args = get_topic_data(w3, method_map, contract_info, log)
         assert args == AttributeDict(expected[i]), [i, method_name]
 
 
@@ -351,11 +304,11 @@ def test_cronos_transfer_source_tokens_with_proxy(ibc):
     w3 = ibc.cronos.w3
     start = w3.eth.get_block_number()
     amount, contract = cronos_transfer_source_tokens_with_proxy(ibc)
-    logs = get_logs(w3, start)
+    logs = get_logs_since(w3, CONTRACT, start)
     escrow = get_escrow_address(cli, channel)
     dst_adr = ibc.chainmain.cosmos_cli().address("signer2")
     expected = get_transfer_source_tokens_topics(dst_adr, amount, contract, escrow)
     assert len(logs) == len(expected)
     for i, log in enumerate(logs):
-        method_name, args = get_topic_data(w3, log)
+        method_name, args = get_topic_data(w3, method_map, contract_info, log)
         assert args == AttributeDict(expected[i]), [i, method_name]