+++ x/icaauth/keeper/keeper.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	"context"
 	"fmt"
 	"time"
 
@@ -10,7 +11,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	capabilitykeeper "github.com/cosmos/cosmos-sdk/x/capability/keeper"
 	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
-	"github.com/cosmos/gogoproto/proto"
 	icacontrollerkeeper "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/controller/keeper"
 	icatypes "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/types"
 	"github.com/crypto-org-chain/cronos/v2/x/icaauth/types"
@@ -46,53 +46,72 @@ func NewKeeper(
 	}
 }
 
-// DoSubmitTx submits a transaction to the host chain on behalf of interchain account
-func (k *Keeper) DoSubmitTx(ctx sdk.Context, connectionID, owner string, msgs []proto.Message, timeoutDuration time.Duration) error {
-	portID, err := icatypes.NewControllerPortID(owner)
+// SubmitTx submits a transaction to the host chain on behalf of interchain account
+func (k *Keeper) SubmitTx(goCtx context.Context, msg *types.MsgSubmitTx) (*types.MsgSubmitTxResponse, error) {
+	msgs, err := msg.GetMessages()
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	data, err := icatypes.SerializeCosmosTx(k.cdc, msgs)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	packetData := icatypes.InterchainAccountPacketData{
 		Type: icatypes.EXECUTE_TX,
 		Data: data,
 	}
-
-	// timeoutDuration should be constraited by MinTimeoutDuration parameter.
-	timeoutTimestamp := ctx.BlockTime().Add(timeoutDuration).UnixNano()
-
-	_, err = k.icaControllerKeeper.SendTx(ctx, nil, connectionID, portID, packetData, uint64(timeoutTimestamp)) //nolint:staticcheck
-	if err != nil {
-		return err
-	}
-
-	return nil
+	return k.SubmitTxWithArgs(goCtx, msg.Owner, msg.ConnectionId, *msg.TimeoutDuration, packetData)
 }
 
-// RegisterInterchainAccount registers an interchain account with the given `connectionId` and `owner` on the host chain
-func (k *Keeper) RegisterInterchainAccount(ctx sdk.Context, connectionID, owner, version string) error {
-	return k.icaControllerKeeper.RegisterInterchainAccount(ctx, connectionID, owner, version)
-}
-
-// GetInterchainAccountAddress fetches the interchain account address for given `connectionId` and `owner`
-func (k *Keeper) GetInterchainAccountAddress(ctx sdk.Context, connectionID, owner string) (string, error) {
+func (k *Keeper) SubmitTxWithArgs(goCtx context.Context, owner, connectionId string, timeoutDuration time.Duration, packetData icatypes.InterchainAccountPacketData) (*types.MsgSubmitTxResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
 	portID, err := icatypes.NewControllerPortID(owner)
 	if err != nil {
-		return "", status.Errorf(codes.InvalidArgument, "invalid owner address: %s", err)
+		return nil, err
+	}
+	minTimeoutDuration := k.MinTimeoutDuration(ctx)
+	// timeoutDuration should be constraited by MinTimeoutDuration parameter.
+	timeoutTimestamp := ctx.BlockTime().Add(
+		types.MsgSubmitTx{
+			TimeoutDuration: &timeoutDuration,
+		}.CalculateTimeoutDuration(minTimeoutDuration)).UnixNano()
+	res, err := k.icaControllerKeeper.SendTx(ctx, nil, connectionId, portID, packetData, uint64(timeoutTimestamp)) //nolint:staticcheck
+	if err != nil {
+		return nil, err
+	}
+	return &types.MsgSubmitTxResponse{
+		Sequence: res,
+	}, nil
+}
+
+// RegisterAccount registers an interchain account with the given `connectionId` and `owner` on the host chain
+func (k *Keeper) RegisterAccount(goCtx context.Context, msg *types.MsgRegisterAccount) (*types.MsgRegisterAccountResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	if err := k.icaControllerKeeper.RegisterInterchainAccount(ctx, msg.ConnectionId, msg.Owner, msg.Version); err != nil {
+		return nil, err
+	}
+	return &types.MsgRegisterAccountResponse{}, nil
+}
+
+// InterchainAccountAddress fetches the interchain account address for given `connectionId` and `owner`
+func (k Keeper) InterchainAccountAddress(goCtx context.Context, req *types.QueryInterchainAccountAddressRequest) (*types.QueryInterchainAccountAddressResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	portID, err := icatypes.NewControllerPortID(req.Owner)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "invalid owner address: %s", err)
 	}
 
-	icaAddress, found := k.icaControllerKeeper.GetInterchainAccountAddress(ctx, connectionID, portID)
+	icaAddress, found := k.icaControllerKeeper.GetInterchainAccountAddress(ctx, req.ConnectionId, portID)
 
 	if !found {
-		return "", status.Errorf(codes.NotFound, "could not find account")
+		return nil, status.Errorf(codes.NotFound, "could not find account")
 	}
 
-	return icaAddress, nil
+	return &types.QueryInterchainAccountAddressResponse{
+		InterchainAccountAddress: icaAddress,
+	}, nil
 }
 
 // ClaimCapability claims the channel capability passed via the OnOpenChanInit callback